Recoil attempt
New recoil implemented, but commented out due to fact that it moves too quickly to be seen, efforts to slow it down are in vain, as increasing timers only cause FPS drops down to as low as 15 FPS.
If anyone can fix this, be my guest, I've stayed up for 2 dyas straight trying to fix it, to no avail. Particles were attempted too, but I couldn't translate form the directX to TL-engine.